src/requirements.txt
python-rtmidi
prompt_toolkit
